Create symlink from host-path to container-path to enhance compatibility
If a project is built outside of the container and the build artifacts
contain absolute paths, this can cause issues as the folder "moves" between
build and analysis time due to the Echidna action running inside a Docker
container. This creates a symlink from the old path to the new one to
try and avoid any issues that might arise from this path change.